refactor(sample): extract the serial profile receiver into its own file
Pure code motion: rx_buf_t, the profile_received semaphore and serial_cb()
move from main.c to profile_serial.c/.h so the receive path can be compiled
into a native_sim test binary (main.c cannot: its main() collides with
ztest's, and the modem headers it includes do not resolve off-target).

One deliberate change rides along: serial_cb() now uses the device pointer
the UART driver passes to the callback instead of reaching for the
file-scope uart_dev global. Same device at runtime -- it is the one the
callback was registered on -- but the function no longer hard-codes uart0,
which is what makes it drivable by a uart-emul test device.
